services/bot/requirements.txt
# Telegram Bot для запуска Mini App
python-telegram-bot[webhooks]==21.7
